Squashed merge from 1003-rel-to-sql (Julian Hyde) Conflicts: 	core/src/main/java/org/apache/calcite/rel/rel2sql/RelToSqlConverter.java 	core/src/test/java/org/apache/calcite/rel/rel2sql/RelToSqlConverterTest.java
