+++ avatica/src/main/java/org/apache/calcite/avatica/util/Cursor.java
@@ -18,7 +18,6 @@ package org.apache.calcite.avatica.util;
 
 import org.apache.calcite.avatica.ColumnMetaData;
 
-import java.io.Closeable;
 import java.io.InputStream;
 import java.io.Reader;
 import java.math.BigDecimal;
@@ -41,7 +40,7 @@ import java.util.Map;
  * Interface to an iteration that is similar to, and can easily support,
  * a JDBC {@link java.sql.ResultSet}, but is simpler to implement.
  */
-public interface Cursor extends Closeable {
+public interface Cursor extends AutoCloseable {
   /**
    * Creates a list of accessors, one per column.
    *
