+++ avatica/src/main/java/org/apache/calcite/avatica/util/IteratorCursor.java
@@ -16,8 +16,6 @@
  */
 package org.apache.calcite.avatica.util;
 
-import java.io.Closeable;
-import java.io.IOException;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
@@ -57,10 +55,12 @@ public abstract class IteratorCursor<E> extends PositionedCursor<E> {
   public void close() {
     current = null;
     position = Position.CLOSED;
-    if (iterator instanceof Closeable) {
+    if (iterator instanceof AutoCloseable) {
       try {
-        ((Closeable) iterator).close();
-      } catch (IOException e) {
+        ((AutoCloseable) iterator).close();
+      } catch (RuntimeException e) {
+        throw e;
+      } catch (Exception e) {
         throw new RuntimeException(e);
       }
     }
